+++ back-end/Public/graphene_hepler.py
@@ -14,17 +14,3 @@ class GrapheneMutation(graphene.Mutation):
 
     def mutate(self, *args, **kw):
         super().mutate(*args, **kw)
-
-
-class GrapheneDjangoObjectType(DjangoObjectType):
-    class Meta:
-        abstract = True
-
-    code = graphene.String(default_value='0')
-    msg = graphene.String()
-
-    def __setattr__(self, key, value):
-        if key == 'code':
-            msg = response_code.get(value, None)
-            self.msg = msg
-        super().__setattr__(key, value)
