+++ back-end/thtr/urls.py
@@ -17,9 +17,10 @@ from django.contrib import admin
 from django.urls import path
 from django.conf.urls import url
 from graphene_django.views import GraphQLView
+from django.views.decorators.csrf import csrf_exempt
 
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    url(r'^graphql', GraphQLView.as_view(graphiql=True))
+    url(r'^graphql/graphiql', csrf_exempt(GraphQLView.as_view(graphiql=True))),
 ]
