+++ deploy/Dockerfiles/nginx/nginx.conf
@@ -32,6 +32,7 @@ http {
         server_name thtr.icu;
 
         root /opt/src/thtr/front-end/build;
+        error_page 405 =200 $uri;
 
         location ~* \.(?:manifest|appcache|html?|xml|json)$ {
             expires    -1s;
