+++ front-end/src/Components/client.js
@@ -1,0 +1,8 @@
+import ApolloClient from 'apollo-boost';
+
+
+const client = new ApolloClient({
+  uri: 'http://192.168.56.235/graphql/graphiql'
+});
+
+export default client;
